+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features:
 
 * Favourite factions. You can now mark factions as favourites (♥ / ♡), and there are new Discord posting options to post only your favourites, or only systems that contain your favourites.
+* There is a new option to use your Colonisation plan names instead of system names in BGS/TW Discord reports. This defaults to `Off`.
 
 ### Changes: